pyproject.toml
[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | migrations/
  | venv
  | _build
  | manage.py
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
skip = [".venv"]

[tool.poetry]
name = "apps"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"        # Updated from <3.10 to support Python 3.11 (KTD-10436)
Django = "3.2.25"
djangorestframework = "3.12.4"
arrow = "0.17.*"
boto = "2.49.*"
boto3 = "^1.42.60"  # Atualizado para suportar PackageType: Image e EphemeralStorage (SURI-134)
Collectfast = "2.2.0"
fnc = "0.5.*"
geojson = "3.1.*"                               # Abandonware (Aug 2019) (Py3.7)
geopy = "1.23.*"
googlemaps = "3.0.*"                            # (Py3.8)
mapbox = "0.18.*"                               # Updated from 0.16.* for Python 3.11 compatibility (collections.abc import fix)
openpyxl = "3.0.9"                              # Going to 3.0.10 causes issues when importing images from Excel files
pyproj = "3.6.*"                                # Updated from 3.2.* for Python 3.11 compatibility (Cython compilation fixes)
shapely = "1.8.*"
tqdm = "4.66.*"                                 # Updated from 4.64.* for Zappa 0.58.x compatibility
zappa = "^0.61.4"                               # Updated from 0.58.* for Python 3.13/3.14 support and SnapStart availability (SURI-136)
sentry-sdk = "2.19.*"
comtele_sdk = "1.1.*"
"pdfminer.six" = "20201018"
PyMuPDF = "1.23.*"                              # Updated from 1.18.* for Python 3.11 compatibility (prebuilt wheels available)
django-bulk-update = "2.2.*" # Abandonware (Aug 2017) (Py3.4)
django-cors-headers = "4.3.1"
django-filter = "23.5"
djangorestframework-gis = "1.0" # False positive for not supported (supports 3.9)
drf-yasg = "1.21.7"
django-sequences = "2.6.*"
django-crispy-forms = "1.14.*"
django-fieldsignals = "0.7.0"
aws-xray-sdk = "2.10.*"
django-simple-history = "3.6.0"
django-rest-passwordreset = "1.4.2" # Watchlist (Oct 2021) (Py3.8)
django-extensions = "3.2.3"
django-premailer = "0.2.*" # Abandonware (Jun 2016) (Py3.5)
dj-database-url = "0.5.*"
django-extra-fields = "3.0.2" # Abandonware (Sep 2020) (Py3.8)
django-ses = "0.8.*"
djangorestframework-jsonapi = "4.3.0"
drf-nested-routers = "0.93.*"
python-decouple = "3.1.*"                       # (Py3.6)
Pillow = "9.1.0"
psycopg2-binary = "2.9.*"                       # Updated to 2.9.* for Python 3.11 compatibility
json_logic_qubit = "0.9.*"
pysaml2 = "7.4.*"
protobuf = "3.20.*"                             # Going over 3.20 will cause errors https://stackoverflow.com/a/72493690
xmlsec = "1.3.*"
setuptools = "57.5.*"
flake8 = { version = "5.0.*", optional = true }
isort = { version = "5.11.*", optional = true }
python-dotenv = "0.20.*"
django-storages = "1.14.6"
werkzeug = "2.2.3"
drf-jwt = "^1.19.2"
django-redis = "5.2.0"
google-auth = "^2.38.0"
django-ratelimit = "^4.1.0"
amplitude-analytics = "1.1.*"

[tool.poetry.group.server]
optional = true

[tool.poetry.group.server.dependencies]
gunicorn = "^21.2.0"

[tool.poetry.group.data-processing]
optional = true

[tool.poetry.group.data-processing.dependencies]
pyarrow = "14.0.2"

[tool.poetry.extras]
ci-tools = ["black", "flake8", "isort", "setuptools"]

[tool.poetry.dev-dependencies]
django-silk = "^5.1.0"
ipdb = "0.13.*" # Watchlist (Jul 2021) (Py3.8)
black = "22.6.*"
pre-commit = "1.21.*"
flake8 = "5.0.*"
isort = "5.11.*"
pytest = "7.4.*"                                # Updated from 5.4.* for Python 3.11 compatibility
pytest-cov = "4.1.*"                            # Updated from 2.12.* for Python 3.11 compatibility
pytest-xdist = "3.5.*"                          # Updated from 1.34.* for Python 3.11 compatibility
pytest-json-report = "1.5.*"
pytest-django = "4.5.*"                         # Updated from 3.10.* for Python 3.11 compatibility
streamlit = "==1.27.*"
pandas = "2.0.3"
numpy = "1.24.3"
geopandas = "0.13.2"
matplotlib = "==3.7.*"
plotly = "^5.24.1"
pypdf2 = "^3.0.1"
beautifulsoup4 = "^4.12.3"
xlsxwriter = "^3.2.0"
scikit-learn = "1.3.2"
scipy = "1.10.1"
regex = "^2024.11.6"
fiona = "==1.8.*"


[build-system]
requires = ["poetry-cor~1.0.0"]
build-backend = "poetry.core.masonry.api"
